requirements.txt
typing
requests
python-dotenv
streamlit